+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added export format **great-expectations**: `datacontract export --format great-expectations`  
 - Added gRPC support to OpenTelemetry integration for publishing test results
 - Added Databricks SQL dialect for `datacontract export --format sql`
+- Added AVRO import support for namespace (#121)
 
 ### Fixed
 
 - Use `sql_type_converter` to build checks.
+- Fixed AVRO import when doc is missing (#121)
 
 ## [0.9.7] - 2024-03-15